+++ assets/src/scripts/monitoring-location/selectors/discrete-data-selector.js
@@ -1,6 +1,6 @@
 import {createSelector} from 'reselect';
 
-import {getCurrentParmCd} from './time-series-selector';
+import {getCurrentVariableID} from './time-series-selector';
 
 /*
  * Returns selector function which returns all of the groundwater levels.
@@ -11,6 +11,19 @@ export const getAllGroundwaterLevels =
     (state) => state.discreteData.groundwaterLevels ? state.discreteData.groundwaterLevels : null;
 
 /*
+ * Returns a selector function which returns an array of all variables that have
+ * groundwater levels
+ * @return {Function} The Function returns an array of the variable properties for each variable
+ * in groundwaterLevels.
+ */
+export const getAllGroundwaterLevelVariables = createSelector(
+    getAllGroundwaterLevels,
+    (gwLevels) => {
+        return gwLevels ? Object.values(gwLevels).map(level => level.variable) : [];
+    }
+);
+
+/*
  * Returns selector function which returns the groundwater levels for the current IV variable
  * @return {Function} - The function returns an Object with the following properties:
  *      @prop {Object} variable
@@ -18,8 +31,8 @@ export const getAllGroundwaterLevels =
  */
 export const getIVCurrentVariableGroundwaterLevels = createSelector(
     getAllGroundwaterLevels,
-    getCurrentParmCd,
-    (gwLevels, parameterCode) => {
-        return gwLevels && parameterCode && gwLevels[parameterCode] ? gwLevels[parameterCode] : {};
+    getCurrentVariableID,
+    (gwLevels, variableID) => {
+        return gwLevels && variableID && gwLevels[variableID] ? gwLevels[variableID] : {};
     }
 );