+++ assets/src/scripts/monitoring-location/selectors/time-series-selector.js
@@ -42,11 +42,20 @@ export const getCustomTimeRange = state => state.ivTimeSeriesState.customIVTimeR
 export const getTimeSeries = state => state.ivTimeSeriesData.timeSeries ? state.ivTimeSeriesData.timeSeries : {};
 
 
-
 export const hasAnyTimeSeries = createSelector(
     getTimeSeries,
     (timeSeries) => {
-        return Object.keys(timeSeries).length ? true : false;
+        return !!Object.keys(timeSeries).length ;
+    }
+);
+
+/*
+ * Selector function which returns a function which returns True if the state contains IV variables
+ */
+export const hasAnyVariables = createSelector(
+    getVariables,
+    (variables) => {
+        return !!(variables && Object.keys(variables).length);
     }
 );